Update ruff crates
Summary:
I updated ruff_annotate_snippets to pick up a bug fix, so might as well update the remaining ruff crates to the same commit.

Expressions now have a node_index field, which means that the field needs to be accounted for when pattern matching, we need to create it (via `AtomicNodeIndex::dummy()`) when constructing expressions, and structures that store expressions are a little larger.

Most of the ruff crates also had to be moved from the "Linux + macOS + Windows" section of third-party/rust/Cargo.toml to the "Linux/aarch64 + Linux/x86_64 + macOS + Windows + Wasm" section, due to them picking a dependency on get-size2, which requires 64-bit atomics.

Reviewed By: stroxler

Differential Revision: D77903268

fbshipit-source-id: c2111fa9f3e78bf56be84c1cb3475fd1c2a28123