Add failing testcase for missing attr base
Summary:
A usecase from pytorch repo:

  ERROR TODO: Expr::attr_infer_for_type attribute base undefined for type: Self (trying to access name) [missing-attribute]
   --> /data/users/migeedz/pytorch/torch/_inductor/pattern_matcher.py:523:59
    |
  523 |         return super().pattern_eq(other) and self.name == other.name
    |                                                           ^^^^^^^^^^

------------------------------

how should we solve this?

It seems we have to handle special forms in `as_attribute_base_no_union`. Specifically | `Type::SpecialForm(SelfType)`

however, I'm noticing that this type doesn't carry any type info that I can use to synthesize an attr base.

I synthesized an `any` and it fixed the error, but can we do better?

It seems that `assert_type(other.outputs, Any)` is wrong, since the annotation is available and it should be `list[Any]`

Reviewed By: stroxler

Differential Revision: D77966256

fbshipit-source-id: 6dee42c32184e7ff000e8caf634f29712e9ff5cc