Check expectations in all user supplied modules
Summary:
Previously we checked for `# E` and errors in the main module. Now we also check it in all of the imports, meaning we can declare errors in `.pyi` files and have them checked (which I want to do).

Requires fixing up a few tests, and unfortunately disabling some non-deterministic tests.

Reviewed By: stroxler

Differential Revision: D78003171

fbshipit-source-id: ce012300b919e3f18a37ba22495f55359f7152c2