Rename `defined_by` to `declared_by`
Summary:
I think if we try to simplify ClassField with an enum, we might
need to be a little clear in our terminology about defined vs declared

In Python, the term "defined" means the name is actually defined, a concept we currently
capture with the "initialization" flags, but that terminology doesn't
really work well for enum cases.

The term "declared" is more typical - not specifically in Python but across
programming languages - for something that is either defined or just declared
to have a type, and I think it therefore better fits the use case here, where
a bare anntation counts.

Reviewed By: ndmitchell

Differential Revision: D78027804

fbshipit-source-id: 7aa6c9b11ea5af2cc07ae84c14ea1e5b991a4323