merge union of all enum member into enum class during simplification
Summary:
There is a conformance test for this in enums_expansion.py, but we already pass it by coincidence due to contextual typing on list literals

The plumbing was a bit weird here, I don't like passing the callback around. Maybe we need something akin to TypeOrder to hold these callbacks?

Reviewed By: ndmitchell

Differential Revision: D77955281

fbshipit-source-id: e71c2659f3fc8aaa4fed633d6c9a39bd65c587a3