Rework some of the naming / commenting on class fields
Summary:
There was an oddly named variable in the class bindings layer, which
was trying to warn us that figuring out what is a class field can be
tricky because the assignments are a bit ambiguous.

But this isn't a good place to highlight that fact, it really should be
a note in the `ClassBinding` doc comments (which is where I put it),
`fields` is a perfectly good name for this variable.

Reviewed By: yangdanny97

Differential Revision: D78031953

fbshipit-source-id: 9c2d3bed068e506a1348c031948dbb7f625b343b