Make T <: None | T
Summary:
The change in D77885729 fixed some things, but broke others. Try and balance all our choices here.

The basic principle is `A | B <: C | D` we must split the LHS first. But for Quantified we can have a bound that wraps a union, and so we still need to split it first.

Reviewed By: stroxler

Differential Revision: D78086121

fbshipit-source-id: 0b1025d12115b6c464f216de7af393768b83c328