Make helpers to manage FlowInfo updates
Summary:
Still dealing with the fact that it took me ~5m to figure out
what default is doing, because the logic was spread out too much
to just read it end-to-end.

If we have helpers on `FlowInfo` to handle create and update, then
it becomes a lot easier to see the meaning of the data structure because
the operations are defined right next to the datatype, instead of hundreds
of lines later.

Reviewed By: rchen152

Differential Revision: D78095851

fbshipit-source-id: 58cbb56c7e3caa3186efb0c299de1dc8c839fceb