Fix inconsistent context for named exprs
Summary:
Before this diff, we were inferring the type of a named expression (walrus operator) twice, once when solving the NamedAssign binding corresponding to the Definition key and once when solving the binding containing the expression.

Furthermore, we were using a different contextual type for each of the inferences. For the definition, we used the annotation on the bound name (if any). For the containing expression, we used the context surrounding the named assignment.

Ideally, we would be able to use context from both sources, but this diff only uses the context from the bound name. Using the context from the surrounding expression would require re-plumbing the bindings a bit, so the NamedAssign binding could depend on the containing expression.

This does mean that we no longer accept code like `xs: list[A] = (ys := [B()])`. This isn't ideal, but is at least as good as mypy, which also rejects this code. Pyright accepts it.

Reviewed By: stroxler

Differential Revision: D78108245

fbshipit-source-id: 6d0340004ab833f8f4e92d3ae7eeb53b3673bd90