use type() for narrowing
Summary:
closes https://github.com/facebook/pyrefly/issues/335

`type(X) == Y` technically means "X is exactly Y and not any subclass of Y" while `type(X) != Y` means "X can be a subclass of Y".

We can't model this in the type system, so at least we can narrow X to Y in the former case (exactly the same as isinstance), and do nothing in the latter case.

Reviewed By: ndmitchell, stroxler

Differential Revision: D78114089

fbshipit-source-id: 9351274439f699a946d3d98fc28b5c404b1ba454