fix indent error in python snippet
Summary: While I was updating the docs, I noticed a syntax error due to an indent bug in the docs, fixing this here. Based on how we are formatting the code snippets, a double indent is needed for the first indent in code snippets to getting the indent to show.

Reviewed By: connernilsen

Differential Revision: D78116695

fbshipit-source-id: 1d5216a8f0af6f767182d574389bf46dfefe05a8