Validate subtype relation for `TypeIs` (#659)
Summary:
(This diff depends on https://github.com/facebook/pyrefly/pull/656)

According to the spec for `TypeIs`
https://typing.python.org/en/latest/spec/narrowing.html#typeis,
the return type in `TypeIs` must be a subtype of the input type being
narrowed. This diff adds a check to that effect for the required
subtyping relation, and adds appropriate test cases.

After this change, all the conformance tests for typeis should pass.

Pull Request resolved: https://github.com/facebook/pyrefly/pull/659

Reviewed By: stroxler

Differential Revision: D78163176

Pulled By: yangdanny97

fbshipit-source-id: 903fd3448a6024ddc1c17229b6f80970339cec85