Cannot assign annotation to variable multiple times
Summary:
tracking the issue here https://github.com/facebook/pyrefly/issues/658

It looks like we're not allowing annotations to the same variable multiple times. The annotations are consistent, so this seems like a reasonable program.

We should probably edit the logic so that if the annotations are consistent, we don't raise the error. Also improve the existing error message when the types aren't consistent. For example, pyright has "Declaration "_x" is obscured by a declaration of the same name "

I will look into that.

Reviewed By: ndmitchell

Differential Revision: D78126380

fbshipit-source-id: 8bf14a01ea90a1e5e42872d73dadf3d5429a907b