Add testcase for dict literal
Summary:
We have already patched PyTorch for dict literals and only have 29 more errors left.

https://www.internalfb.com/intern/everpaste/?handle=GOlppx5QOKgpBE8EAIYwjMYBnK4TbsIXAAAB&phabricator_paste_number=1865869914

those are the dicts that need to be created programatically though, which we should fix.

Issue:
https://github.com/facebook/pyrefly/issues/602

I'm not 100% sure if this unittest covers the remaining cases in this paste. Nonetheless, we should fix it. If it doesn't I'll look more into it.

will start by looking into expr_infer to see if there's any special casing needed for dict.

Reviewed By: ndmitchell

Differential Revision: D78118720

fbshipit-source-id: 33e36a7290048bdf1d2760d9be8c97325d3459a4