Smarter way to find directories to propose to enable typechecking
Summary:
Update logic to fetch the directories to show for enabling typechecking. The goal is to maximize then number of directories shown that will have between 10 to 100 type errors, to help the user get started with typechecking.

The algorithm loops through each child layer until we find the first layer that produces 10 directories/files between 10-100 errors, OR if that doesn't exist, return the layer that produced the most directories/files which satisfies that condition. Ties will be broken using the layer closest to the root layer

Reviewed By: connernilsen

Differential Revision: D77319328

fbshipit-source-id: 6a551695af09279840ed31c5a9cf512648cde6b8