move print logic into init.rs
Summary: Title, I realized that since we need to add logic to prompt the user to select directories, we need to expose the directories shown to `init.rs`. As a result, we should handling printing in `init.rs` as well, and this will make the individual logic of fetching the directories easier to test as well.

Reviewed By: MaggieMoss

Differential Revision: D78036446

fbshipit-source-id: ca23a8ff3b1579ebb57e5f9d8ba314053018069a