write pyrefly typeshed to disk instead of contentsasuri
Summary:
pyrefly's source of truth for typeshed is in-memory. because of this, in order to see these files in the IDE, we have used VSCode's documentcontentprovider API (D74087183). But this has a few drawbacks:
- this does not work on non-vscode (vim/emacs/other extensions)
- these files are hard to deal with code navigation on (not real files)

To solve both of these problems, we can just write this directory to disk. This diff does this once per lsp session (should be enough). If we can't write to /tmp/, go to definition will not work on these files.

## Other approaches
I considered a few different approaches in order to get code navigation working on these files:
- bundle the full tree with vscode extension (this only fixes it for vscode and we need to store a mapping)
- bundle the full tree with pyrefly binary (this is not feasable)
- keep using contentsasuri but try getting pyrefly to accept it D77962628, [better approach](https://fburl.com/diff/a3tmoqin) (this only fixes it for vscode and is flaky)

The approach taken in this diff seems like the best option. The only drawback is access to tempdir: if we can't write to it we can't show these definitions.

Reviewed By: ndmitchell, grievejia

Differential Revision: D78184707

fbshipit-source-id: b374d3cd06362f9436d03f8a5b745bab41602374