Modify AnnAssign logic to handle inconsistent annotations
Summary:
Instead of directly rejecting redefinitions of annotated assignments, why not delegate that to the existing BindingExpect?

We went from 16 invalid annotations to 3 invalid annotations in PyTorch with this change

https://www.internalfb.com/intern/everpaste/?handle=GL6bxR5o8aNFY6oEALvUWSVOUV5xbsIXAAAz&phabricator_paste_number=1867426958

vs

https://www.internalfb.com/intern/everpaste/?handle=GJDL2R4GGE96ELUCAOlcyV-MNmc9bsIXAAAz&phabricator_paste_number=1867424040

note that the 13 issues we eliminated were due to this specific issue (do we did not remove any other "invalid annotations" errors. The remaining 3 are a separate issue.

One existing testcase changed:

  from typing import assert_type, Literal, Final
  class A:
    x: int
    y: str
    def __init__(self):
        self.x: Literal[1] = 1   // this is ok in mypy
        self.y: Final = "y"  // this is ok in pyright
  def f(a: A):
    assert_type(a.x, int)

But this seems okay, since accepting both is also reasonable (seeing that they are accepted in mypy and pyright and make no conformance change)

Fixes issue https://github.com/facebook/pyrefly/issues/658

Reviewed By: stroxler

Differential Revision: D78195958

fbshipit-source-id: a13ffeee36196ad8eec835b9b3e415808f7fb0c6