Simplify variance inference
Summary:
As far as I can tell (and as far as the tests can tell) there was a lot of code to go from a QName to a Class. A lot of that code was fairly weird (it checked the ancestors of a different class to see if they happened to match) followed by a generic case where it looked up an export. All of those thing are fairly expensive.

Happily, a few diffs ago we made it just have the class it was operating on. And now we don't need to look it up. We also get to remove a bunch of `continue`'s that were pretty dubious and marked as "should be a panic".

Reviewed By: stroxler

Differential Revision: D78213960

fbshipit-source-id: f1d3f8baa084f4686a7fbf7ce690adcb8290997b