Add test cases clarifying Callable attribute behavior
Summary:
We implement behavior whereby a callable assigned on a class body is assumed to
be a custom callable type not implementing `__get__`, but we lack a test case
for this.

This logic is actually not compatible with Pyright and mypy; it is approximately
in line with how they treat "normal" attributes because they assume instance
shadowing might occur - in which case no method binding would happen and access
on an instance would actually be the bare callable - but they actually do assume
`__get__` in cases where method binding unambiguously has a chance to occur.

I added a second test case, where Pyrefly currently behaves the same, with ClassVar
in play because this is the case that highlights what Pyright and mypy are actually
doing and why.

----------

We probably need to fix this, there are a number of example codebases that actually
rely on `ClassVar[Callable[...]]` assigned in function bodies, but I'm not sure yet
how to approach it because there's a closely related question of how to deal with
callback protocols (which have the same issue, and it matters because lots of
internal decorator libraries use them!)

I posted about this in
https://discuss.python.org/t/when-should-we-assume-callable-types-are-method-descriptors/92938
and will probably give at least a couple days to see if anyone in OSS has opinions
before I chart a path forward.

In the meantime, it's good to have these issues documented in unit tests.

Reviewed By: grievejia

Differential Revision: D78121013

fbshipit-source-id: ca60ada21214e6186128eb3c1b489eba83c7934a