Visiting Expressions to generate additional predicates
Summary: For some predicates, I need to also visit the expressions. Additionally, for the predicate CalleeToCaller, I also need to keep track of the `container` which is the closest function or class containing the expressions (see D78173041). That's why I am visiting expressions for the statement attributes that don't get visit when traversing only the body. Open to suggestions about better ways to traverse the ast while keeping the info I need.

Reviewed By: ndmitchell

Differential Revision: D78173040

fbshipit-source-id: 20d944ae2e33be975b15825acb5a495a1e5f3ba5