Tweak the `type(...)` as base class example(s)
Summary:
The issue of dealing with `type(...)` in a base class list originally came
up in the context of a specific metaclass issue that has been solved already.

But the general problem remains: sometimes - particularly when dealing with
metaclasses in C extensions - it might be necessary to use `type(...)` to
get at a class that isn't otherwise easily accessible in a base class because
of the way the runtime works.

We ought to try to support it. After D78120610, which landed last week,
we actually do handle it in some cases but the support is not as good as
it should be.

This diff splits the test into three tests, one of which is actually passing
because of D78120610 and the other two are failing because the types
coming out of `type(...)` still aren't as precise as they could be.

Reviewed By: ndmitchell

Differential Revision: D78246480

fbshipit-source-id: 7421193f70db4f3d5836a5279aed4d42468c98e3