Change the semantics of class objects initialized on class bodies
Summary:
Fixes https://github.com/facebook/pyrefly/issues/645 (although it exposes some
other issues)

It isn't 100% clear how to handle a class field initialized on the class body whose
type is `type[X]`. We know from some existing examples that `X` needs to be treated
covariantly, because plenty of codebases rely on nested classes and assume covariant
overrides.

Just to head off any suggestion that user code should be changed rather than us
fixing this in the type checker: note that at the moment, there is no way to
explicitly do this in the type system; `Final` would make the field
non-overridable... this is one reason why PEP 767 (extending ReadOnly from just
typed dict to normal classes) would be useful. But that PEP is still a draft at
this time.

In addition, class objects themselves don't have an annotation, and this change
is assuming class defs are implicitly read-only which I'm pretty confident is
the right choice; I pretty strongly feel that if we do that, we also need to make
it easy to override a nested class *def* with a plain assignment as in some
of the test cases here.

---------

Something odd in this diff: it turns out `read_only_reason` is being ignored
by the actual attribute write checks, we've duplicated the read-onlyness checks in
inconsistent ways.

In fact, it appears to me that we *never* respect read-onlyness on class attributes
today. I'll attempt to partially address this in a fast-follow.

But I also am confinced this diff, and the surprising behavior, is solid evidence
that the current one-struct, "bag of flags" structure of ClassField isn't working
well and we really need to switch to enums that more clearly indicate the case
analysis (there's a TODO in the code that's been present for months for me to do
this; as more logic has been added the refactor has become both more difficult and
more necessary).

Reviewed By: ndmitchell

Differential Revision: D78241403

fbshipit-source-id: 4f730f11a9fc9951167f4dff18f429566ea6da17