Prototype of a pysa export option
Summary:
# Context
We want Pysa to use type information from Pyrefly, since that will become the standard type checker at Meta, and because it provides better type information.

To achieve this, we need Pyrefly to provide all the information Pysa needs.
We thought about multiple approaches to do this:
* Run Pyrefly once, dump all the information needed to files in some format (say, JSON). Let Pysa parse all these files at startup.
* Run Pyrefly as a server, and query it from Pysa, i.e Inter Process Communication (IPC).
* Run Pyrefly and Pysa in the same process, which would require Rust and OCaml interop.

Each of these solutions have different tradeoffs (implementation speed, maintainance burden, runtime speed, etc.).

# This diff
This diff is an experimental approach to the first solution: dump all the information we need to files, since this is definitely the easiest approach. This is not meant to be used in Production, but it will serve as a way to find exactly all the information Pysa needs.

Reviewed By: stroxler

Differential Revision: D78022303

fbshipit-source-id: 3fe7a1be89c6ad48bd470f03082d128e5b4d38b0