Implement the solve for `Binding::SelfTypeLiteral`
Summary:
Note that our current implementation of `Self` does not support TypedDict; the
current datatype only allows normal class instance types to live inside Self. This
is probably fine, just good to be aware of; we could relax it by making `Self` take
an arbitrary `Type` as its bound (which would be more consistent with the wording
of PEP 673) but it's not yet clear to me that there's a use case where we need this
and at any rate any such change is certainly out of scope for this bugfix.

Reviewed By: rchen152

Differential Revision: D78344294

fbshipit-source-id: 268f2762091f4a815a6b92449e93b73a6fb5d856