AnswersSolver-side logic for direct use of `typing.Self`
Summary:
This diff completes my current effort on the Self type, which I realized never
really worked at all prior to this diff - the *only* correctly-handled use of the
self type was implicit annotations of the first parameter of a function, any other
use was just the special form (which is a useless type).

On the answers solver side, we can completely bypass normal type analysis on the
`typing.Self` annotation - there's really no point running normal analysis because
it makes good sense to first check for a `Binding::SelfTypeLiteral` since that's
very cheap (just a hashed text range).

This fixes https://github.com/facebook/pyrefly/issues/649, the
name `typing.Self` now binds correctly (I only belatedly realized that it was
wrong in all cases before, the fact that we can't distinguish a proper Self from
the special form in our printing means every single test using `Self` literals
was actually not behaving as the author expected - it worked in exactly zero
cases before this diff).

-----

That said, the tests are still clearly exposing several problems:
- The fact that we don't print any difference between the special form Self
  and a bound Self is pretty problematic. We need to decide a way to indicate
  this much more clearly!
- We aren't actually implementing the right semantics. `Self` should be treated
  like a type variable to the extent possible, and we are not doing this. I'm
  unsure exactly how many semantics are broken due to this, but pretty much
  all solver behaviors are wrong, plus attribute lookups don't work.

I've marked both tests as still buggy because there's a fair amount of more
theoretical work to do. That said, I suspect the logic is now good enough
for a pretty decent fraction of uses of `Self`, which was not the case before,

Reviewed By: rchen152

Differential Revision: D78344295

fbshipit-source-id: 36ae7df55989221233db32025f471a36a0951781