Support jump to definition on hover
Summary:
We can keep track of a set of displayed class names in `TypeDisplayContext` while we are printing types, and later use to add additional hover content to help user to directly jump to definition.

I encoded the exact location of the symbol in the URL in the fragment part, like `file://foo.py#L1,2`. This is non-standard and only recognized by VSCode. We could gate it by client, but even if it's not supported by other editors, it can at least jump to the right file.

Reviewed By: kinto0

Differential Revision: D78296086

fbshipit-source-id: 9968095aec10bb1f5c6867b1ffa104b6520ef47b