Add needs_triage.yml GitHub workflow
Summary:
Adds workflow to identify issues and PRs that have gone more than 48 hours without triage.

* For issues, "without triage" means that the issue has not yet had a type (issue/bug/feature) applied and is not labeled as "question" (which means we need more info from the reporter).
* For PRs, "without triage" means that the PR does not have a requested reviewer, assignee, or review.

I tested some, but not all, of the pieces of this workflow in a personal fork, but I couldn't test things like issue types, which seem to be set at an organization level.

My confidence that this works as expected is therefore very low; I've commented out the actual labeling of issues and PRs as "needs-triage" until I can test this. Testing requires a copy of the workflow to be checked in on main.

I did at least run this through a JS syntax checker.

Reviewed By: migeed-z

Differential Revision: D78374027

fbshipit-source-id: 39f7765fcc195ce27a62373a3404e0931a0e27ed