Change make_bound_method to take a type
Summary:
All existing callers pass an Instance, but in the next diff I want to pass a class, for binding metaclass __call__. This updates the existing callers to pass in a Type instead, used in the next diff.

I considered extending the Instance variant, but it didn't feel right. But happy for input on an alternate approach.

Reviewed By: stroxler

Differential Revision: D78430633

fbshipit-source-id: 52a9234b6b0edcc469f0ce917757b3070a62938b