Drop a note on first-use pinning in `AssignToAttribute`
Summary:
I realized as I was doing a deep dive into `BindingClassField` that we
are duplicating expression here (similar to how we duplicate them in
narrowing).

This really worried me for a few minutes, because duplicated expressions are
potentially very difficult to analyze deterministically if placeholder types
(like an empty list) are in play.

But after some thought I realized that unlike narrowing - where it is
impractical to try to force a clear ordering because of how the binding graph
works - contextual typing of attribute assignment means this is actually okay,
because we always have to analyze the ClassField binding before the assignment
in order to contextually type assignments.

This feels important enough to actually have a comment making a note about it;
I opted to put it on `AssignToAttribute`'s solve since that's where the contextual
typing winds up saving us.

Reviewed By: ndmitchell

Differential Revision: D78418304

fbshipit-source-id: 90b8ba343d022d948a3ea8617b6a5328fa844318