Move `MethodThatSetsAttr`
Summary:
I want to split BindingClassField for clearer case analysis, as the first step
of doing more careful case analysis throughout the attribute-analysis stack.

One of the factors both contributing to why I want to do this and getting in my
way is duplication in the existing data model - things like "is this attribute
defined in a method" are modeled in ways that are pretty difficult to track
because of duplication and the logic being spread across many modules.

Before I dig into that refactor, it's helpful to consolidate some types that I'll
need to juggle when deduplicating and converting to case analysis.

Reviewed By: ndmitchell

Differential Revision: D78464142

fbshipit-source-id: 2cfdeb26d5a7c770de235affd6f4b2a5eb26fed3