Inline `fields_from_metadata` and avoid crashing
Summary:
shipfastcrash

We've observed multiple IDE crashes in practice caused by an `unwrap()` inside `fields_from_metadata()`. The function basically asserts that its parameter (aka the metadata) must contain a non-None typeddict metadata field, and crash if the assertion doesn't hold.

This kind of assertion is generally unsafe to make under our current incremental check mechanism. During incremental check, we might be working with an outdated solution and hence outdated class type from other modules, but with an updated metadata for the class. That could happen, and when it happens we would be 99.99% sure that we are working with outdated data, after which it doesn't really matter what we do -- whatever result we compute from then on it's going to be outdated anyway and we'll be recompute shortly afterwards. But we shouldn't crash since there won't be any recompute once we crash.

This diff makes it such that `fields_from_metadata()` no longer asserts for non-None typeddict metadata. Callers of it just can return some dummy value when they hit `None`. I've also inlined the function into callers given that there were just 2 of them and they use slightly different dummy values.

Reviewed By: ndmitchell

Differential Revision: D78455189

fbshipit-source-id: 4558ecce4eb165d1491f8b1f4a50778a0f6fe704