if _value_ is inherited from enum.Enum, use enum members to infer a type
Summary:
`enum.Enum` annotates `_value_` as Any, so when accessing `.value` on an instance of the enum class (not a literal member type) that did not provide an explicit annotation for `_value_` we get `Any`.

We can infer a better type by using the raw value types assigned to each member, but we should take care to use the promoted values to avoid creating huge unions.

closes https://github.com/facebook/pyrefly/issues/695

Reviewed By: stroxler

Differential Revision: D78498605

fbshipit-source-id: 91101e44860537beed54693276ffed4d2d57f618