Create helper function for creating a ConfigFile from existing mypy/pyright/pyproject configs
Summary: We want to add functionality to be able to run `pyrefly check` without `pyrefly init` by reading the existing mypy/pyright/pyproject configs directly. This diff takes the first step by creating a helper function for loading config so that we can reuse this during pyrefly check if we can't find an existing config.

Reviewed By: connernilsen

Differential Revision: D78488916

fbshipit-source-id: 9953289a3627c54b82f3275b797bf4d94087340f