Prep for making BindingClassField use an enum for cases
Summary:
I have concluded from a combination of my experiences trying to read and understand
class field logic as well as an accumulation of bugs in our github issues that the
class field code really needs to be using case analysis rather than just a record
with a ton of flags that interact in ways that are challenging to track.

I want to start refactoring to use enums from the inside out, beginning with the
binding layer. This diff prepares for that, by leaving the inherent class field
information (the name and class) directly on the BindingClassField but moving
most of the other data about the field definition into an enum.

This forces some complexity in the `calculate_class_field` code because that
function - which is very long, very complex, and part of my motivation for this
refactor - is currently quite coupled to having a ton of fields which we now have
to extract from the enum.

The enum is degenerate right now, I just wanted to get the blocking refactor done
first; I'll begin breaking it into cases next.

Reviewed By: grievejia

Differential Revision: D78464145

fbshipit-source-id: aea8aa44b042bcf6576f5cf67273446218ac7f89