Enumerate the cases clearly at binding time
Summary:
I finally have the bindings looking the way I would like.

There's still a chunk of tech debt around this because `RawClassFieldInitialValue`
is being used to express (badly) the actual cases reachable in various places; I
hope to clean that up next.

But the bindings themselves express the cases I'd like them to at this stage.

-----

I was a little sad to need a DeclaredWithoutAnnotation case - this is unreachable
in normal classes, but I was unable to model `collections.namedtuple` synthesized
class semantics without it (fields behave a lot like a field with annotation `Any`,
but because of the way bindings work it's very difficult to actually construct
a synthetic annotation for them, it's too easy to hit a duplicate range and panic).

Reviewed By: grievejia

Differential Revision: D78464147

fbshipit-source-id: e32814ad76fd7e2f97aa6e263097820e930cc8e4