Simplify synthesized field logic just a bit
Summary:
This funciton is big, there are a lot of variables in scope.

We can at least slightly reduce the noise by avoiding an unnecessary
`key_field` var and an unnecessary `idx_for_promise` - the insert
binding operation gives us everything we need.

While we're at it, we can also rename the annotation `annotation` to
cut down on line noise just a bit.

Reviewed By: grievejia

Differential Revision: D78464151

fbshipit-source-id: 6acfdd674ac69ae5d3f806d0263a90df4cb60465