Deduplicate method name
Summary:
A lot of minimum-edit-distance diffs for shipping various features led to us
recording the method name twice in method-defined class fields.

This costs RAM and was also harder to read or refactor. This diff deduplicates:
we can just store `MethodThatSetsAttr`, which knows everything we need about
the method name, in the `RawClassFieldInitialization`. (Note that the reason we
have that boolean flag at all is that we precompute whether it's a "recognized"
method because the mechanism for saving attribute writes needs to know to
prefer recognized methods to unrecognized ones when tracking the first
assignment)

Reviewed By: grievejia

Differential Revision: D78464155

fbshipit-source-id: 7d780a2e046f151bbe509c97ed0351863ee71f1d