Deduplicate match on `initial_value`
Summary:
The initial value match is pretty hairy - for historical reasons,
we are relying on invariants that aren't enforcible in the current
type structure.

I'd like to make it clearer before proceeding with refactors, but
there's an even simpler thing in my way which is the double-match
on `initial_value`.

Let's simplify first; this leads to a bit of code duplication (two
uses of `Binding::Forward` but I think having a single match is
simpler in a more important structural way).

Reviewed By: grievejia

Differential Revision: D78464143

fbshipit-source-id: a624fb000b7fa5bb07770cf1ef4ebf70bef99f00