Split defined-in-method vs defined-in-class-body
Summary:
This diff continues splitting the critical different cases that are straightforward
to distinguish at bindings time, which I hope will help me make better sense of
the logic at `calculate_class_field` time later.

Here, I split the cases where the field is defined in only methods versus the
class body. These cases are possible to distinguish today via the (duplicated)
information content of `implicit_def_method` and `initial_value`, but I find
it considerably harder to actually track how these two fields get used and
how they are defined versus a clearly named case.

This also should make it more straightforward to extend the data model later
if we want to to track multiple assignments to class fields, which is something
we've considered doing for better type inference.

-------

I realized at this stage that really, `DefinedInClassBody` deserves a clearer
case analysis around *how* it's defined; the existing `RawClassFieldInitialization`
enum kinda-sorta models this but in a way that keeps confusing me.

I think that most likely the `MethodLike` case of `ClassFieldDefinition` should
really be moved into a sub-case of `DefinedInClassBody`.

I decided for this diff to leave `RawClassFieldInitialization` enum as it is
even thought that's a bit messy because I think it makes sense to keep fiddling
with the case analysis and wait until the dust settles before rewriting the
low-level helper types.

Reviewed By: grievejia

Differential Revision: D78464153

fbshipit-source-id: a3b0addd09ec1e3e207ce58c132a04becea2f719