Use a single match for value / initial_value
Summary:
I'm about to refactor to get rid of the fact that these two fields don't
clearly express what's going on, and they duplicate information in ways that
makes downstream logic harder to read.

It's a lot easier to do the deduplication of we use a single match to clearly
express the duplicated logic in one pass first.

Reviewed By: grievejia

Differential Revision: D78464141

fbshipit-source-id: dcc9d105674fdb3f5242a5098bd01bdb9ff2d710