Use a hack to clarify `RawClassFieldInitialization`
Summary:
I'm still not ready to try to clean this type up, but it's a
bit of a mess right now because it's showing up in places where
some of the variants don't make sense.

At this stage in my refactor, it's really useful to have an explicit
match on all cases and explicitly call out that one of the cases
is unreachable. I'll clean it up a bit later (the type needs to be
replaced with something else, but I really want to do one more case
split first)

Reviewed By: grievejia

Differential Revision: D78464140

fbshipit-source-id: 12e8ddefcb3a3ed5f22ffac626d084a3474edb14