Split method vs other cases for ClassFieldDefinition
Summary:
This diff begins case-analyzing class field binding-time information, by splitting
"method-like" class fields coming from `def` statements from other fields.

This is useful because:
- defs aren't easy to model in the existing structure, since in many cases they
  behave "as if" they were annotated but they lack an acutal annotation; an enum
  lets us make this explicit at binding time
- neither annotation nor value are really needed for method modeling, and we can
  represent this much more cleanly in the bindings (there's a cost downstream
  for now since the ClassField logic isn't refactored, but that should get better
  later).
- it allows us to eliminate the `is_function_without_return_annotation` field
  from the non-method-like `Simple` case

We do pay a cost for this in the short run in `calculate_class_field`; the block we
already had to spit out all the old record fields now also has to use some heap
storage hooks so that they can synthesize dummy values as refs (this is a very rust
problem, the synthesized values are owned and are defined inside a branch of the
logic, so without doing something to get heap storage, we can't produce a ref and
we don't want to introduce clones).

I expect this downstream logic to eventually be simplified, but I need to start
somewhere and for now, as I'm modeling the bindings more clearly, we have to pay
a cost somewhere.

Reviewed By: grievejia

Differential Revision: D78464148

fbshipit-source-id: a1acf0c1a60905cd8c6a78d0c6a28e9a755440be