Disable typechecking in repos other than the selected directory
Summary: Following the ability to suppress errors on Pyrefly Init for repos with a large amount of errors, we will be disabling typechecking on everything other than the selected directory. This will enable the user to have the ability for a clean typecheck. The idea of this is rather than a new dev using pyrefly, seeing thousands of errors, and giving up on Pyrefly right away, we will let them try out Pyrefly on some directories that they select, and if they end up enjoying using it, they can expand it to more and more of their repo.

Reviewed By: connernilsen

Differential Revision: D78436634

fbshipit-source-id: d09f7e62e21af46915703bc1d5c3620e3de41ce5