Disallow replacing builtins as typing.Any
Summary:
dluo noticed that when running Pyrefly on Fixit with `replace_imports_with_any = ["*"]`, he gets a TON of type errors (D78513502 for internal viewers).

I realized earlier that the fix here would just be to disallow replacing `builtins` with any, since it's in prelude but we likely still need to import it for stuff, and when testing locally, that fixed the issue.

Reviewed By: ndmitchell

Differential Revision: D78535289

fbshipit-source-id: fb56d5fdf0f11ee47d3482670a0335329a077fe9