Refactoring to address gaps in variable facts
Summary:
We have a few facts that are common for all declarations, so, instead of repeating the same code 5 times (for classes, function, import, variable and module), I am extracting the duplicated code and creating the struct DeclarationInfo that is used to create DeclarationLocation, DefinitionLocation, and ContainingTopLevelDeclaration. Doing this, automatically fixes the gap for the ContainingTopLevelDeclaration that was not previously added for variables.
Additionally, I am also adding the container for variables

Reviewed By: ndmitchell

Differential Revision: D78484124

fbshipit-source-id: 91e27ada0956efca539a5085adff209333eb86b2